pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "langchain-core>=0.3.72",
    "langchain-groq>=0.3.6",
    "langgraph>=0.6.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "ruff>=0.12.5",
    "uvicorn>=0.35.0",
    "yfinance>=0.2.65",
]
